cleanup indexes create fields, fix url options tests
