--- packages/compass-indexes/src/components/create-index-fields/index.ts
@@ -1,2 +1,0 @@
-import CreateIndexFields from './create-index-fields';
-export default CreateIndexFields;
